Move YamlUserConfiguration to authority module. (#24513)
* Move YamlUserConfiguration to authority module.

* Fix import.

* Change org.junit.Test to org.junit.jupiter.api.Test
