+++ kernel/authority/core/src/test/java/org/apache/shardingsphere/authority/yaml/swapper/YamlEngineUserConfigurationTest.java
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.metadata.user.yaml.swapper;
+package org.apache.shardingsphere.authority.yaml.swapper;
 
-import org.apache.shardingsphere.infra.metadata.user.yaml.config.YamlUserConfiguration;
+import org.apache.shardingsphere.authority.yaml.config.YamlUserConfiguration;
 import org.apache.shardingsphere.infra.util.yaml.YamlEngine;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.util.Properties;
